requirements.txt
langchain>=0.2
langchain-core>=0.2
langchain-openai>=0.2
langgraph>=0.2
faiss-cpu
pydantic>=2
python-dotenv
fastapi
uvicorn
tiktoken
rapidfuzz
pypinyin
pytest
pypdf
python-docx
pillow
pytest-dotenv